requirements.txt
certifi==2019.9.11
chardet==3.0.4
future==0.18.0
idna==2.8
oauthlib==3.1.0
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.6
